improved drag experience + moved utils
