+++ .vscode/settings.json
@@ -14,6 +14,7 @@
 		"pickable",
 		"popovertarget",
 		"Registrator",
+		"stylesheet",
 		"svgs",
 		"templating",
 		"tinymce",
