+++ e2e/upgrader.spec.ts
@@ -1,5 +1,5 @@
 const { rest } = window.MockServiceWorker;
-import { umbracoPath } from '@umbraco-cms/backoffice/utils';
+import { umbracoPath } from 'src/packages/core/utils/index.js';
 import {
 	ProblemDetails,
 	RuntimeLevelModel,
