+++ package.json
@@ -14,70 +14,71 @@
 		"./localization-api": "./dist-cms/libs/localization-api/index.js",
 		"./observable-api": "./dist-cms/libs/observable-api/index.js",
 		"./auth": "./dist-cms/shared/auth/index.js",
-		"./event": "./dist-cms/packages/core/event/index.js",
-		"./lit-element": "./dist-cms/packages/core/lit-element/index.js",
 		"./icon": "./dist-cms/shared/icon-registry/index.js",
 		"./models": "./dist-cms/shared/models/index.js",
 		"./resources": "./dist-cms/shared/resources/index.js",
 		"./router": "./dist-cms/shared/router/index.js",
 		"./style": "./dist-cms/shared/style/index.js",
-		"./utils": "./dist-cms/shared/utils/index.js",
 		"./action": "./dist-cms/packages/core/action/index.js",
+		"./audit-log": "./dist-cms/packages/audit-log/index.js",
+		"./block": "./dist-cms/packages/block/index.js",
+		"./code-editor": "./dist-cms/packages/templating/code-editor/index.js",
 		"./collection": "./dist-cms/packages/core/collection/index.js",
 		"./components": "./dist-cms/packages/core/components/index.js",
 		"./content-type": "./dist-cms/packages/core/content-type/index.js",
 		"./culture": "./dist-cms/packages/core/culture/index.js",
+		"./current-user": "./dist-cms/packages/user/current-user/index.js",
 		"./data-type": "./dist-cms/packages/core/data-type/index.js",
 		"./debug": "./dist-cms/packages/core/debug/index.js",
+		"./dictionary": "./dist-cms/packages/dictionary/index.js",
+		"./document-blueprint": "./dist-cms/packages/documents/document-blueprints/index.js",
+		"./document-type": "./dist-cms/packages/documents/document-types/index.js",
+		"./document": "./dist-cms/packages/documents/documents/index.js",
+		"./dynamic-root": "./dist-cms/packages/dynamic-root/index.js",
 		"./entity-action": "./dist-cms/packages/core/entity-action/index.js",
 		"./entity-bulk-action": "./dist-cms/packages/core/entity-bulk-action/index.js",
+		"./event": "./dist-cms/packages/core/event/index.js",
 		"./extension-registry": "./dist-cms/packages/core/extension-registry/index.js",
-		"./server-file-system": "./dist-cms/packages/core/server-file-system/index.js",
 		"./id": "./dist-cms/packages/core/id/index.js",
+		"./language": "./dist-cms/packages/language/index.js",
+		"./lit-element": "./dist-cms/packages/core/lit-element/index.js",
 		"./localization": "./dist-cms/packages/core/localization/index.js",
+		"./log-viewer": "./dist-cms/packages/log-viewer/index.js",
+		"./media-type": "./dist-cms/packages/media/media-types/index.js",
+		"./media": "./dist-cms/packages/media/media/index.js",
+		"./member-group": "./dist-cms/packages/members/member-group/index.js",
+		"./member-type": "./dist-cms/packages/members/member-type/index.js",
+		"./member": "./dist-cms/packages/members/member/index.js",
 		"./menu": "./dist-cms/packages/core/menu/index.js",
 		"./modal": "./dist-cms/packages/core/modal/index.js",
 		"./notification": "./dist-cms/packages/core/notification/index.js",
+		"./object-type": "./dist-cms/packages/object-type/index.js",
+		"./package": "./dist-cms/packages/packages/package/index.js",
+		"./partial-view": "./dist-cms/packages/templating/partial-views/index.js",
 		"./picker-input": "./dist-cms/packages/core/picker-input/index.js",
-		"./property": "./dist-cms/packages/core/property/index.js",
 		"./property-action": "./dist-cms/packages/core/property-action/index.js",
 		"./property-editor": "./dist-cms/packages/core/property-editor/index.js",
-		"./section": "./dist-cms/packages/core/section/index.js",
-		"./sorter": "./dist-cms/packages/core/sorter/index.js",
-		"./store": "./dist-cms/packages/core/store/index.js",
-		"./themes": "./dist-cms/packages/core/themes/index.js",
-		"./tree": "./dist-cms/packages/core/tree/index.js",
-		"./variant": "./dist-cms/packages/core/variant/index.js",
-		"./workspace": "./dist-cms/packages/core/workspace/index.js",
-		"./repository": "./dist-cms/packages/core/repository/index.js",
-		"./temporary-file": "./dist-cms/packages/core/temporary-file/index.js",
-		"./block": "./dist-cms/packages/block/index.js",
-		"./audit-log": "./dist-cms/packages/audit-log/index.js",
-		"./dictionary": "./dist-cms/packages/dictionary/index.js",
-		"./document": "./dist-cms/packages/documents/documents/index.js",
-		"./document-blueprint": "./dist-cms/packages/documents/document-blueprints/index.js",
-		"./document-type": "./dist-cms/packages/documents/document-types/index.js",
-		"./media": "./dist-cms/packages/media/media/index.js",
-		"./media-type": "./dist-cms/packages/media/media-types/index.js",
-		"./member": "./dist-cms/packages/members/member/index.js",
-		"./member-group": "./dist-cms/packages/members/member-group/index.js",
-		"./member-type": "./dist-cms/packages/members/member-type/index.js",
-		"./package": "./dist-cms/packages/packages/package/index.js",
-		"./language": "./dist-cms/packages/language/index.js",
-		"./dynamic-root": "./dist-cms/packages/dynamic-root/index.js",
-		"./log-viewer": "./dist-cms/packages/log-viewer/index.js",
+		"./property": "./dist-cms/packages/core/property/index.js",
 		"./relation-type": "./dist-cms/packages/relations/relation-types/index.js",
 		"./relations": "./dist-cms/packages/relations/relations/index.js",
-		"./tags": "./dist-cms/packages/tags/index.js",
+		"./repository": "./dist-cms/packages/core/repository/index.js",
+		"./section": "./dist-cms/packages/core/section/index.js",
+		"./server-file-system": "./dist-cms/packages/core/server-file-system/index.js",
+		"./sorter": "./dist-cms/packages/core/sorter/index.js",
 		"./static-file": "./dist-cms/packages/static-file/index.js",
-		"./partial-view": "./dist-cms/packages/templating/partial-views/index.js",
+		"./store": "./dist-cms/packages/core/store/index.js",
 		"./stylesheet": "./dist-cms/packages/templating/stylesheets/index.js",
+		"./tags": "./dist-cms/packages/tags/index.js",
 		"./template": "./dist-cms/packages/templating/templates/index.js",
+		"./temporary-file": "./dist-cms/packages/core/temporary-file/index.js",
+		"./themes": "./dist-cms/packages/core/themes/index.js",
+		"./tree": "./dist-cms/packages/core/tree/index.js",
 		"./user-group": "./dist-cms/packages/user/user-group/index.js",
-		"./current-user": "./dist-cms/packages/user/current-user/index.js",
-		"./user": "./dist-cms/packages/user/user/index.js",
 		"./user-permission": "./dist-cms/packages/user/user-permission/index.js",
-		"./code-editor": "./dist-cms/packages/templating/code-editor/index.js",
+		"./user": "./dist-cms/packages/user/user/index.js",
+		"./utils": "./dist-cms/packages/core/utils/index.js",
+		"./variant": "./dist-cms/packages/core/variant/index.js",
+		"./workspace": "./dist-cms/packages/core/workspace/index.js",
 		"./external/backend-api": "./dist-cms/external/backend-api/index.js",
 		"./external/dompurify": "./dist-cms/external/dompurify/index.js",
 		"./external/lit": "./dist-cms/external/lit/index.js",
