+++ src/packages/block/block-grid/components/block-grid-entries/block-grid-entries.element.ts
@@ -1,3 +1,8 @@
+import {
+	getAccumulatedValueOfIndex,
+	getInterpolatedIndexOfPositionInWeightMap,
+	isWithinRect,
+} from '@umbraco-cms/backoffice/utils';
 import { UmbBlockGridEntriesContext } from '../../context/block-grid-entries.context.js';
 import type { UmbBlockGridEntryElement } from '../block-grid-entry/index.js';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
@@ -11,62 +16,14 @@ import {
 	type resolveVerticalDirectionArgs,
 } from '@umbraco-cms/backoffice/sorter';
 
-// Utils:
-// TODO: Move these methods into their own files:
-
-function expandRect(rect: DOMRect, verticalAdd: number, horizontalAdd: number) {
-	return new DOMRect(
-		rect.left - horizontalAdd,
-		rect.top - verticalAdd,
-		rect.width + verticalAdd * 2,
-		rect.height + horizontalAdd * 2,
-	);
-}
-
-function getInterpolatedIndexOfPositionInWeightMap(target: number, weights: Array<number>) {
-	const map = [0];
-	weights.reduce((a, b, i) => {
-		return (map[i + 1] = a + b);
-	}, 0);
-	const foundValue = map.reduce((a, b) => {
-		const aDiff = Math.abs(a - target);
-		const bDiff = Math.abs(b - target);
-
-		if (aDiff === bDiff) {
-			return a < b ? a : b;
-		} else {
-			return bDiff < aDiff ? b : a;
-		}
-	});
-	const foundIndex = map.indexOf(foundValue);
-	const targetDiff = target - foundValue;
-	let interpolatedIndex = foundIndex;
-	if (targetDiff < 0 && foundIndex === 0) {
-		// Don't adjust.
-	} else if (targetDiff > 0 && foundIndex === map.length - 1) {
-		// Don't adjust.
-	} else {
-		const foundInterpolationWeight = weights[targetDiff >= 0 ? foundIndex : foundIndex - 1];
-		interpolatedIndex += foundInterpolationWeight === 0 ? interpolatedIndex : targetDiff / foundInterpolationWeight;
-	}
-	return interpolatedIndex;
-}
-
-function getAccumulatedValueOfIndex(index: number, weights: Array<number>) {
-	const len = Math.min(index, weights.length);
-	let i = 0,
-		calc = 0;
-	while (i < len) {
-		calc += weights[i++];
-	}
-	return calc;
-}
-
+/**
+ * Notice this utility method is not really shareable with others as it also takes areas into account. [NL]
+ */
 function resolveVerticalDirectionAsGrid(
 	args: resolveVerticalDirectionArgs<UmbBlockGridLayoutModel, UmbBlockGridEntryElement>,
 ) {
-	// If this has areas, we do not want to move:
-	if (args.relatedModel.areas.length > 0 && expandRect(args.relatedRect, 0, 0)) {
+	// If this has areas, we do not want to move, unless we are at the edge
+	if (args.relatedModel.areas.length > 0 && isWithinRect(args.pointerX, args.pointerY, args.relatedRect, -10)) {
 		return null;
 	}
 
@@ -154,6 +111,12 @@ export class UmbBlockGridEntriesElement extends UmbLitElement {
 	//
 	#sorter = new UmbSorterController<UmbBlockGridLayoutModel, UmbBlockGridEntryElement>(this, {
 		...SORTER_CONFIG,
+		onStart: () => {
+			this.#context.onDragStart();
+		},
+		onEnd: () => {
+			this.#context.onDragEnd();
+		},
 		onChange: ({ model }) => {
 			this.#context.setLayouts(model);
 		},
@@ -295,13 +258,18 @@ export class UmbBlockGridEntriesElement extends UmbLitElement {
 
 			uui-button-group {
 				padding-top: 1px;
-				display: grid;
 				grid-template-columns: 1fr auto;
+
+				--umb-block-grid--is-dragging--variable: var(--umb-block-grid--is-dragging) none;
+				display: var(--umb-block-grid--is-dragging--variable, grid);
 			}
 
-			/*.umb-block-grid__layout-container[data-area-length='0'] {
-				min-height: 100px;
-			}*/
+			.umb-block-grid__layout-container[data-area-length='0'] {
+				--umb-block-grid--is-dragging--variable: var(--umb-block-grid--is-dragging) 1;
+				min-height: calc(var(--umb-block-grid--is-dragging--variable, 0) * var(--uui-size-11));
+				border: calc(var(--umb-block-grid--is-dragging--variable, 0) * 1px) dashed var(--uui-color-border);
+				border-radius: var(--uui-border-radius);
+			}
 		`,
 	];
 }
