+++ src/packages/block/block-grid/context/block-grid-entry.context.ts
@@ -1,3 +1,4 @@
+import { closestColumnSpanOption } from '../utils/index.js';
 import { UMB_BLOCK_GRID_MANAGER_CONTEXT } from './block-grid-manager.context.js';
 import { UMB_BLOCK_GRID_ENTRIES_CONTEXT } from './block-grid-entries.context-token.js';
 import {
@@ -17,28 +18,6 @@ import {
 	appendToFrozenArray,
 } from '@umbraco-cms/backoffice/observable-api';
 import { combineLatest } from '@umbraco-cms/backoffice/external/rxjs';
-
-function closestColumnSpanOption(target: number, map: Array<number>, max: number) {
-	if (map.length > 0) {
-		const result = map.reduce((a, b) => {
-			if (a > max) {
-				return b;
-			}
-			const aDiff = Math.abs(a - target);
-			const bDiff = Math.abs(b - target);
-
-			if (aDiff === bDiff) {
-				return a < b ? a : b;
-			} else {
-				return bDiff < aDiff ? b : a;
-			}
-		});
-		if (result) {
-			return result;
-		}
-	}
-	return;
-}
 
 export class UmbBlockGridEntryContext
 	extends UmbBlockEntryContext<
