+++ src/packages/core/property-editor/uis/tiny-mce/plugins/tiny-mce-mediapicker.plugin.ts
@@ -1,12 +1,12 @@
 import type { TinyMcePluginArguments } from '@umbraco-cms/backoffice/components';
 import { UmbTinyMcePluginBase } from '@umbraco-cms/backoffice/components';
-import { UmbMediaHelper } from '@umbraco-cms/backoffice/utils';
 import type { UmbModalManagerContext } from '@umbraco-cms/backoffice/modal';
 import { UMB_MEDIA_TREE_PICKER_MODAL, UMB_MODAL_MANAGER_CONTEXT } from '@umbraco-cms/backoffice/modal';
 import type { UMB_CURRENT_USER_CONTEXT, UmbCurrentUserModel } from '@umbraco-cms/backoffice/current-user';
 import type { RawEditorOptions } from '@umbraco-cms/backoffice/external/tinymce';
 import { UmbTemporaryFileRepository } from '@umbraco-cms/backoffice/temporary-file';
 import { UmbId } from '@umbraco-cms/backoffice/id';
+import { sizeImageInEditor, uploadBlobImages } from '@umbraco-cms/backoffice/media';
 
 interface MediaPickerTargetData {
 	altText?: string;
@@ -26,7 +26,6 @@ interface MediaPickerResultData {
 }
 
 export default class UmbTinyMceMediaPickerPlugin extends UmbTinyMcePluginBase {
-	#mediaHelper: UmbMediaHelper;
 	#currentUser?: UmbCurrentUserModel;
 	#modalContext?: UmbModalManagerContext;
 	#currentUserContext?: typeof UMB_CURRENT_USER_CONTEXT.TYPE;
@@ -35,7 +34,6 @@ export default class UmbTinyMceMediaPickerPlugin extends UmbTinyMcePluginBase {
 	constructor(args: TinyMcePluginArguments) {
 		super(args);
 
-		this.#mediaHelper = new UmbMediaHelper();
 		this.#temporaryFileRepository = new UmbTemporaryFileRepository(args.host);
 
 		this.consumeContext(UMB_MODAL_MANAGER_CONTEXT, (modalContext) => {
@@ -79,7 +77,7 @@ export default class UmbTinyMceMediaPickerPlugin extends UmbTinyMcePluginBase {
 				const content = e.content;
 
 				// Handle images that are pasted in
-				this.#mediaHelper.uploadBlobImages(this.editor, content);
+				uploadBlobImages(this.editor, content);
 			});
 		}
 	}
@@ -203,7 +201,7 @@ export default class UmbTinyMceMediaPickerPlugin extends UmbTinyMcePluginBase {
 
 			// When image is loaded we are ready to call sizeImageInEditor.
 			const onImageLoaded = () => {
-				this.#mediaHelper?.sizeImageInEditor(this.editor, imgElm, img.url);
+				sizeImageInEditor(this.editor, imgElm, img.url);
 				this.editor.dispatch('Change');
 			};
