+++ src/packages/core/utils/pagination-manager/pagination.manager.ts
