+++ src/packages/core/utils/path/path-encode.function.ts
