+++ src/packages/core/utils/path/path-folder-name.function.ts
@@ -1,0 +1,3 @@
+import { generateAlias } from '../string/generate-umbraco-alias.function.js';
+
+export const pathFolderName = generateAlias;
