+++ src/packages/core/utils/selection-manager/selection.manager.test.ts
