+++ src/packages/core/utils/selection-manager/selection.manager.ts
