+++ src/packages/core/utils/string/generate-umbraco-alias.function.ts
