+++ src/packages/core/utils/string/increment-string.function.ts
