+++ src/packages/core/utils/string/split-string-to-array.ts
