+++ src/packages/media/media/index.ts
@@ -1,11 +1,12 @@
 import './components/index.js';
 
-export * from './repository/index.js';
-export * from './workspace/index.js';
-export * from './tracked-reference/index.js';
-export * from './user-permissions/index.js';
 export * from './components/index.js';
 export * from './entity.js';
+export * from './repository/index.js';
+export * from './tracked-reference/index.js';
+export * from './user-permissions/index.js';
+export * from './utils/index.js';
+export * from './workspace/index.js';
 
 export { UMB_MEDIA_TREE_ALIAS } from './tree/index.js';
 export { UMB_MEDIA_COLLECTION_ALIAS } from './collection/index.js';
