--- src/shared/utils/component-has-manifest-property.function.ts
@@ -1,7 +1,0 @@
-import type { ManifestBase } from '@umbraco-cms/backoffice/extension-api';
-
-export function componentHasManifestProperty(
-	component: HTMLElement,
-): component is HTMLElement & { manifest: ManifestBase } {
-	return component ? 'manifest' in component : false;
-}
