--- src/shared/utils/index.ts
@@ -1,20 +1,0 @@
-export * from './component-has-manifest-property.function.js';
-export * from './diff.type.js';
-export * from './ensure-path-ends-with-slash.function.js';
-export * from './generate-umbraco-alias.function.js';
-export * from './increment-string.function.js';
-export * from './media-helper.service.js';
-export * from './pagination-manager/pagination.manager.js';
-export * from './path-decode.function.js';
-export * from './path-encode.function.js';
-export * from './path-folder-name.function.js';
-export * from './selection-manager/selection.manager.js';
-export * from './umbraco-path.function.js';
-export * from './math.js';
-export * from './split-string-to-array.js';
-
-declare global {
-	interface Window {
-		Umbraco: any;
-	}
-}
