--- src/shared/utils/path-folder-name.function.ts
@@ -1,3 +1,0 @@
-import { generateAlias } from './generate-umbraco-alias.function.js';
-
-export const pathFolderName = generateAlias;
